+++ DesignerMacros/src/services-expose-as-http-endpoint/_common/getDefaultRoutePrefix.ts
@@ -1,12 +1,19 @@
 function getDefaultRoutePrefix(includeLastPathSeparator: boolean): string {
-    const defaultApiRoutePrefix = "api";
+    const defaultApiRoutePrefix = "api/";
     const apiSettingsId = "4bd0b4e9-7b53-42a9-bb4a-277abb92a0eb";
 
     let settingsGroup = application.getSettings(apiSettingsId);
     let route = settingsGroup ? settingsGroup.getField("Default API Route Prefix").value : null;
 
-    if (!route) {
+    // if the group is not present, use the default value
+    if(!settingsGroup)
+    {
         route = defaultApiRoutePrefix;
+    }
+    // if the route is null (or set to blank in settings, which results in null)
+    // set it to blank (the actual value in settings)
+    if (!route) {
+        route = "";
     }
 
     if (includeLastPathSeparator && !route.endsWith("/")) {